requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.44
python-dotenv==1.0.1
pydantic==1.10.14
